refactor(codex-adapter): carve event_mapper.rs into session_state/parent_id_sink/turn_lifecycle
Todo #247 task 1: event_mapper.rs was heading toward the 300-line
ceiling as CollabAgent work lands. Split CodexSessionState into
session_state.rs, ParentIdSink into parent_id_sink.rs, and the
turn/plan/usage notification handlers into turn_lifecycle.rs, all
verbatim moves with re-exports so external call sites keep compiling.
