+++ packages/core-rs/crates/mainframe-adapter-codex/src/lib.rs
@@ -26,6 +26,7 @@ pub(crate) mod image_generation_history;
 pub(crate) mod image_generation_render;
 pub mod item_types;
 pub mod jsonrpc;
+pub(crate) mod parent_id_sink;
 pub mod plan_mode_handler;
 pub mod quota_identity;
 pub mod quota_pull;
@@ -33,12 +34,14 @@ pub mod quota_rate_limit;
 pub mod rollout_reader;
 pub(crate) mod rollout_reconstruct;
 pub mod session;
+pub(crate) mod session_state;
 pub(crate) mod thread_item_render;
 pub(crate) mod thread_item_variants;
 pub mod thread_registry;
 pub mod title_generator;
 pub mod transcript;
 pub mod turn_config;
+pub(crate) mod turn_lifecycle;
 pub mod types;
 pub(crate) mod web_search_history;
 pub(crate) mod web_search_render;
